Improved environment variable
Only `TEST` and `PRODUCTION` is supported. Depending on the set string, the `setEnvironment()` function did only set a different API URL.

Instead of using strings for this, a boolean value makes more sense.

In case, if the variable / parameter is not set to true or false, the function `setEnvironment()` will return an error message, that the given value is not valid and it needs to be `true` or `false`.